% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_score_consensus.R
\name{as.data.frame.mutateR_consensus_batch}
\alias{as.data.frame.mutateR_consensus_batch}
\title{Extract comparison table from batch analysis}
\usage{
\method{as.data.frame}{mutateR_consensus_batch}(
  x,
  row.names = NULL,
  optional = FALSE,
  include_correlations = FALSE,
  ...
)
}
\arguments{
\item{x}{A mutateR_consensus_batch object}

\item{include_correlations}{Logical. Include individual pairwise correlations (default FALSE).}
}
\value{
A data frame with one row per gene and summary statistics.
}
\description{
Extract comparison table from batch analysis
}
