% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_batch.R
\name{diagnose_batch_result}
\alias{diagnose_batch_result}
\title{Diagnose batch analysis completeness}
\usage{
diagnose_batch_result(batch_result)
}
\arguments{
\item{batch_result}{A mutateR_consensus_batch object}
}
\value{
Data frame summarising each gene's status
}
\description{
Diagnose batch analysis completeness
}
