% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNA_feature_extraction.R
\name{get_feature_names}
\alias{get_feature_names}
\title{Get list of all available feature names}
\usage{
get_feature_names(include_id = FALSE)
}
\arguments{
\item{include_id}{Logical. Include 'grna_id' column (default FALSE)}
}
\value{
Character vector of feature names
}
\description{
Returns the names of all features computed by \code{extract_grna_features()}.
Useful for feature selection and ablation studies.
}
\examples{
feature_names <- get_feature_names()
length(feature_names)

# Categorize features
positional <- grep("^pos\\\\d+_", feature_names, value = TRUE)
dinuc <- grep("^freq_", feature_names, value = TRUE)

}
