% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_ablation.R
\name{get_intrinsic_features}
\alias{get_intrinsic_features}
\title{Get list of intrinsic recognition features}
\usage{
get_intrinsic_features()
}
\value{
Character vector of intrinsic recognition feature names
}
\description{
Returns feature names related to intrinsic Cas9:gRNA:DNA recognition
and not gRNA transcription.
}
\details{
Intrinsic recognition features include:
\itemize{
\item Seed region nucleotides (positions 12-20)
\item PAM-proximal preferences (positions 17-20)
\item GC content (overall and regional)
\item Thermodynamic properties
\item Self-complementarity
}
}
\examples{
intrinsic <- get_intrinsic_features()
length(intrinsic)

}
\seealso{
\code{\link{get_transcription_features}}
}
