% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_score_consensus.R
\name{print.mutateR_consensus_batch}
\alias{print.mutateR_consensus_batch}
\title{Print method for batch consensus analysis}
\usage{
\method{print}{mutateR_consensus_batch}(x, ...)
}
\description{
Print method for batch consensus analysis
}
