% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_analysis.R
\name{print.mutateR_tiered_shap}
\alias{print.mutateR_tiered_shap}
\title{Print method for tiered SHAP analysis}
\usage{
\method{print}{mutateR_tiered_shap}(x, ...)
}
\arguments{
\item{x}{A mutateR_tiered_shap object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for tiered SHAP analysis
}
