% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multi_score_correl.R
\name{print.score_agreement}
\alias{print.score_agreement}
\title{Print method for score agreement summary}
\usage{
\method{print}{score_agreement}(x, ...)
}
\arguments{
\item{x}{Output from summarize_score_agreement()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for score agreement summary
}
