% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_score_consensus.R
\name{summary.mutateR_consensus_batch}
\alias{summary.mutateR_consensus_batch}
\title{Summary method for batch consensus analysis}
\usage{
\method{summary}{mutateR_consensus_batch}(object, ...)
}
\description{
Summary method for batch consensus analysis
}
