% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_analysis.R
\name{summary.mutateR_shap_analysis}
\alias{summary.mutateR_shap_analysis}
\title{Summary method for SHAP analysis}
\usage{
\method{summary}{mutateR_shap_analysis}(object, n_features = 20, ...)
}
\arguments{
\item{object}{A mutateR_shap_analysis object}

\item{n_features}{Integer. Number of top features to show (default 20)}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for SHAP analysis
}
