% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multi_score_correl.R
\name{summary.score_agreement}
\alias{summary.score_agreement}
\title{Summary method for score agreement}
\usage{
\method{summary}{score_agreement}(object, ...)
}
\description{
Summary method for score agreement
}
